API makes request to UUID API and saves new antibody.
Created template for Mockserver that returns UUID, HuBMAP base ID and
HuBMAP ID according to the authorization header. That allows Pytest to
send the UUID in the request and test it gets saved correctly.

Added Requests library.

API now hits the UUID API to get a new UUID before saving an antibody.
FIX: URL is hard-coded to the mockserver.

Added test for UUID creation.